+++ src/main/java/com/jfrog/mavendeptree/dependenciesresults/MavenDepTreeResults.java
@@ -1,4 +1,4 @@
-package com.jfrog.dependenciesResults;
+package com.jfrog.mavendeptree.dependenciesresults;
 
 import java.util.Map;
