+++ src/main/java/com/jfrog/mavendeptree/dependenciesresults/MavenDependencyNode.java
@@ -1,4 +1,4 @@
-package com.jfrog.dependenciesResults;
+package com.jfrog.mavendeptree.dependenciesresults;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
